move types into root directory
