+++ packages/devextreme/js/__internal/scheduler/appointments/data_provider/m_appointment_data_provider.ts
@@ -1,6 +1,6 @@
 import config from '@js/core/config';
 import { combineRemoteFilter } from '@ts/scheduler/r1/filterting/index';
-import type { AppointmentDataItem, SafeAppointment } from '@ts/scheduler/r1/types';
+import type { AppointmentDataItem, SafeAppointment } from '@ts/scheduler/types';
 import type { AppointmentDataAccessor } from '@ts/scheduler/utils';
 
 import { AppointmentDataSource } from './m_appointment_data_source';
