+++ packages/devextreme/js/__internal/scheduler/appointments/rendering_strategies/m_strategy_base.ts
@@ -3,8 +3,8 @@ import { extend } from '@js/core/utils/extend';
 import { isNumeric, isObject } from '@js/core/utils/type';
 import { current as currentTheme } from '@js/ui/themes';
 import { dateUtilsTs } from '@ts/core/utils/date';
-import type { SafeAppointment } from '@ts/scheduler/r1/types';
 import { isAppointmentTakesAllDay } from '@ts/scheduler/r1/utils/index';
+import type { SafeAppointment } from '@ts/scheduler/types';
 
 import { createAppointmentAdapter } from '../../m_appointment_adapter';
 import timeZoneUtils from '../../m_utils_time_zone';
