+++ packages/devextreme/js/__internal/scheduler/appointments/rendering_strategies/m_strategy_horizontal_month_line.ts
@@ -1,6 +1,6 @@
 import query from '@js/common/data/query';
 import dateUtils from '@js/core/utils/date';
-import type { SafeAppointment } from '@ts/scheduler/r1/types';
+import type { SafeAppointment } from '@ts/scheduler/types';
 
 import { sortAppointmentsByStartDate } from '../data_provider/m_utils';
 import HorizontalAppointmentsStrategy from './m_strategy_horizontal';
