+++ packages/devextreme/js/__internal/scheduler/r1/components/base/group_panel.tsx
@@ -3,7 +3,7 @@ import { createReRenderEffect, InfernoWrapperComponent } from '@ts/core/r1/runti
 import type { RefObject } from '@ts/core/r1/types';
 
 import { VERTICAL_GROUP_ORIENTATION } from '../../../constants';
-import type { Group, GroupOrientation } from '../../types';
+import type { Group, GroupOrientation } from '../../../types';
 import { isVerticalGroupingApplied } from '../../utils/index';
 import type { DefaultProps } from '../types';
 import { GroupPanelHorizontal } from './group_panel_horizontal';
