+++ packages/devextreme/js/__internal/scheduler/r1/components/base/group_panel_horizontal.tsx
@@ -1,6 +1,6 @@
 import { BaseInfernoComponent } from '@ts/core/r1/runtime/inferno/index';
 
-import type { GroupRenderItem } from '../../types';
+import type { GroupRenderItem } from '../../../types';
 import { GroupPanelHorizontalRow } from './group_panel_horizontal_row';
 import type { GroupPanelBaseProps } from './group_panel_props';
 import { GroupPanelBaseDefaultProps } from './group_panel_props';
