+++ packages/devextreme/js/__internal/scheduler/r1/components/base/group_panel_props.ts
@@ -1,7 +1,7 @@
 import type { PropsWithClassName, PropsWithStyles } from '@ts/core/r1/index';
 import type { JSXTemplate } from '@ts/core/r1/types';
 
-import type { GroupItem, GroupPanelData, GroupRenderItem } from '../../types';
+import type { GroupItem, GroupPanelData, GroupRenderItem } from '../../../types';
 import type { DefaultProps, PropsWithViewContext, ResourceCellTemplateProps } from '../types';
 
 export interface GroupPanelBaseProps extends
