+++ packages/devextreme/js/__internal/scheduler/r1/components/base/time_panel_table.tsx
@@ -3,7 +3,7 @@ import { createReRenderEffect, InfernoWrapperComponent } from '@ts/core/r1/runti
 import type { JSXTemplate, RefObject } from '@ts/core/r1/types';
 import { Fragment } from 'inferno';
 
-import type { GroupOrientation, TimePanelData } from '../../types';
+import type { GroupOrientation, TimePanelData } from '../../../types';
 import type { DateTimeCellTemplateProps, DefaultProps, PropsWithViewContext } from '../types';
 import { AllDayPanelTitle } from './all_day_panel_title';
 import { CellBase, CellBaseDefaultProps } from './cell';
