+++ packages/devextreme/js/__internal/scheduler/r1/utils/month.ts
@@ -1,7 +1,7 @@
 import dateLocalization from '@js/common/core/localization/date';
 import dateUtils from '@js/core/utils/date';
 
-import type { CalculateCellIndex } from '../types';
+import type { CalculateCellIndex } from '../../types';
 import {
   getCalculatedFirstDayOfWeek,
   isDateInRange,
