+++ packages/devextreme/js/__internal/scheduler/r1/utils/timeline_month.ts
@@ -1,6 +1,6 @@
 import dateUtils from '@js/core/utils/date';
 
-import type { CalculateStartViewDate } from '../types';
+import type { CalculateStartViewDate } from '../../types';
 import { setOptionHour } from './base';
 import { getViewStartByOptions } from './month';
