+++ packages/devextreme/js/__internal/scheduler/r1/utils/views.ts
@@ -1,7 +1,7 @@
 import { isObject } from '@js/core/utils/type';
 
 import { VIEW_TYPES } from '../../constants';
-import type { ViewType } from '../types';
+import type { ViewType } from '../../types';
 
 export const getCurrentView = (
   currentView: string | ViewType,
